added retry logic to celery. tests are deterministic again :)

+++ backend/tasks/definitions.py
@@ -41,7 +41,8 @@ from backend.logic.stock_data import (
     get_symbols_table,
     fetch_iex_price,
     during_trading_day,
-    get_all_active_symbols
+    get_all_active_symbols,
+    SeleniumDriverError
 )
 from backend.logic.visuals import (
     compile_and_pack_player_sidebar_stats,
@@ -53,13 +54,13 @@ from backend.logic.visuals import (
 )
 from backend.tasks.celery import (
     celery,
-    SqlAlchemyTask,
+    BaseTask,
     pause_return_until_subtask_completion
 )
 from backend.tasks.redis import rds
 
 
-@celery.task(name="async_get_user_information", bind=True, base=SqlAlchemyTask)
+@celery.task(name="async_get_user_information", bind=True, base=BaseTask)
 def async_get_user_information(self, user_id):
     return get_user_information(user_id)
 
@@ -68,14 +69,14 @@ def async_get_user_information(self, user_id):
 # Price fetching and caching #
 # -------------------------- #
 
-@celery.task(name="async_fetch_active_symbol_prices", bind=True, base=SqlAlchemyTask)
+@celery.task(name="async_fetch_active_symbol_prices", bind=True, base=BaseTask)
 def async_fetch_active_symbol_prices(self):
     active_symbols = get_all_active_symbols(db_session)
     for symbol in active_symbols:
         async_fetch_price.delay(symbol)
 
 
-@celery.task(name="async_fetch_price", bind=True, base=SqlAlchemyTask)
+@celery.task(name="async_fetch_price", bind=True, base=BaseTask)
 def async_fetch_price(self, symbol):
     """Whatever method is used to get a price, the return should always be a (price, timestamp) tuple.
     """
@@ -84,7 +85,7 @@ def async_fetch_price(self, symbol):
     return price, timestamp
 
 
-@celery.task(name="async_cache_price", bind=True, base=SqlAlchemyTask)
+@celery.task(name="async_cache_price", bind=True, base=BaseTask)
 def async_cache_price(self, symbol: str, price: float, last_updated: float):
     """We'll store the last-updated price of each monitored stock in redis. In the short-term this will save us some
     unnecessary data API call.
@@ -109,12 +110,12 @@ def async_cache_price(self, symbol: str, price: float, last_updated: float):
 # Game management #
 # --------------- #
 
-@celery.task(name="async_get_user_responses_for_pending_game", bind=True, base=SqlAlchemyTask)
+@celery.task(name="async_get_user_responses_for_pending_game", bind=True, base=BaseTask)
 def async_get_user_responses_for_pending_game(self, game_id):
     return get_user_responses_for_pending_game(game_id)
 
 
-@celery.task(name="async_add_game", bind=True, base=SqlAlchemyTask)
+@celery.task(name="async_add_game", bind=True, base=BaseTask)
 def async_add_game(self, creator_id, title, mode, duration, buy_in, n_rebuys, benchmark, side_bets_perc,
                    side_bets_period, invitees):
     opened_at = time.time()
@@ -141,7 +142,7 @@ def async_add_game(self, creator_id, title, mode, duration, buy_in, n_rebuys, be
     create_game_invites_entries(game_id, creator_id, user_ids, opened_at)
 
 
-@celery.task(name="async_respond_to_game_invite", bind=True, base=SqlAlchemyTask)
+@celery.task(name="async_respond_to_game_invite", bind=True, base=BaseTask)
 def async_respond_to_game_invite(self, game_id, user_id, status):
     assert status in ["joined", "declined"]
     response_time = time.time()
@@ -156,7 +157,7 @@ def async_respond_to_game_invite(self, game_id, user_id, status):
     start_game_if_all_invites_responded(game_id)
 
 
-@celery.task(name="async_service_open_games", bind=True, base=SqlAlchemyTask)
+@celery.task(name="async_service_open_games", bind=True, base=BaseTask)
 def async_service_open_games(self):
     open_game_ids = get_open_game_invite_ids()
     status_list = []
@@ -165,17 +166,17 @@ def async_service_open_games(self):
     pause_return_until_subtask_completion(status_list, "async_service_open_games")
 
 
-@celery.task(name="async_service_one_open_game", bind=True, base=SqlAlchemyTask)
+@celery.task(name="async_service_one_open_game", bind=True, base=BaseTask)
 def async_service_one_open_game(self, game_id):
     service_open_game(game_id)
 
 
-@celery.task(name="async_get_game_info", bind=True, base=SqlAlchemyTask)
+@celery.task(name="async_get_game_info", bind=True, base=BaseTask)
 def async_get_game_info(self, game_id):
     return get_game_info(game_id)
 
 
-@celery.task(name="async_get_game_info_for_user", bind=True, base=SqlAlchemyTask)
+@celery.task(name="async_get_game_info_for_user", bind=True, base=BaseTask)
 def async_get_game_info_for_user(self, user_id):
     return get_game_info_for_user(user_id)
 
@@ -185,7 +186,7 @@ def async_get_game_info_for_user(self, user_id):
 # ---------------- #
 
 
-@celery.task(name="async_suggest_symbols", base=SqlAlchemyTask)
+@celery.task(name="async_suggest_symbols", base=BaseTask)
 def async_suggest_symbols(text):
     with db_session.connection() as conn:
         to_match = f"{text.upper()}%"
@@ -197,23 +198,23 @@ def async_suggest_symbols(text):
     return [{"symbol": entry[1], "label": f"{entry[1]} ({entry[2]})"} for entry in symbol_suggestions]
 
 
-@celery.task(name="async_update_symbols_table", bind=True, default_retry_delay=10, base=SqlAlchemyTask)
-def async_update_symbols_table(self):
-    try:
-        symbols_table = get_symbols_table()
-        print("writing to db...")
-        with db_session.connection() as conn:
-            conn.execute("TRUNCATE TABLE symbols;")
-            db_session.remove()
+@celery.task(name="async_update_symbols_table", bind=True, default_retry_delay=10, base=BaseTask)
+def async_update_symbols_table(self, n_rows=None):
+    symbols_table = get_symbols_table(n_rows)
+    if symbols_table.empty:
+        raise SeleniumDriverError
 
-        with db_session.connection() as conn:
-            symbols_table.to_sql("symbols", conn, if_exists="append", index=False)
-            db_session.commit()
-    except Exception as exc:
-        raise self.retry(exc=exc)
+    print("writing to db...")
+    with db_session.connection() as conn:
+        conn.execute("TRUNCATE TABLE symbols;")
+        db_session.remove()
+
+    with db_session.connection() as conn:
+        symbols_table.to_sql("symbols", conn, if_exists="append", index=False)
+        db_session.commit()
 
 
-@celery.task(name="async_place_order", base=SqlAlchemyTask)
+@celery.task(name="async_place_order", base=BaseTask)
 def async_place_order(user_id, game_id, symbol, buy_or_sell, order_type, quantity_type, market_price, amount,
                       time_in_force, stop_limit_price=None):
     """Placing an order involves several layers of conditional logic: is this is a buy or sell order? Stop, limit, or
@@ -230,7 +231,7 @@ def async_place_order(user_id, game_id, symbol, buy_or_sell, order_type, quantit
                 market_price, float(amount), time_in_force, stop_limit_price)
 
 
-@celery.task(name="async_process_single_order", base=SqlAlchemyTask)
+@celery.task(name="async_process_single_order", base=BaseTask)
 def async_process_single_order(order_id):
     timestamp = time.time()
     if get_order_expiration_status(order_id):
@@ -249,7 +250,7 @@ def async_process_single_order(order_id):
                   order_ticket["order_type"], order_ticket["price"], market_price, order_ticket["quantity"], timestamp)
 
 
-@celery.task(name="async_process_all_open_orders", bind=True, base=SqlAlchemyTask)
+@celery.task(name="async_process_all_open_orders", bind=True, base=BaseTask)
 def async_process_all_open_orders(self):
     """Scheduled to update all orders across all games throughout the trading day
     """
@@ -265,7 +266,7 @@ def async_process_all_open_orders(self):
 # ------- #
 
 
-@celery.task(name="async_invite_friend", bind=True, base=SqlAlchemyTask)
+@celery.task(name="async_invite_friend", bind=True, base=BaseTask)
 def async_invite_friend(self, requester_id, invited_username):
     """Since the user is sending the request, we'll know their user ID via their web token. We don't post this
     information to the frontend for other users, though, so we'll look up their ID based on username
@@ -275,7 +276,7 @@ def async_invite_friend(self, requester_id, invited_username):
     table_updater(friends, requester_id=requester_id, invited_id=invited_id, status="invited", timestamp=time.time())
 
 
-@celery.task(name="async_respond_to_friend_invite", bind=True, base=SqlAlchemyTask)
+@celery.task(name="async_respond_to_friend_invite", bind=True, base=BaseTask)
 def async_respond_to_friend_invite(self, requester_username, invited_id, decision):
     """Since the user is responding to the request, we'll know their user ID via their web token. We don't post this
     information to the frontend for other users, though, so we'll look up the request ID based on the username
@@ -285,18 +286,18 @@ def async_respond_to_friend_invite(self, requester_username, invited_id, decisio
     table_updater(friends, requester_id=requester_id, invited_id=invited_id, status=decision, timestamp=time.time())
 
 
-@celery.task(name="async_get_friends_details", bind=True, base=SqlAlchemyTask)
+@celery.task(name="async_get_friends_details", bind=True, base=BaseTask)
 def async_get_friends_details(self, user_id):
     friend_ids = get_friend_ids(user_id)
     return get_user_details_from_ids(friend_ids)
 
 
-@celery.task(name="async_suggest_friends", bind=True, base=SqlAlchemyTask)
+@celery.task(name="async_suggest_friends", bind=True, base=BaseTask)
 def async_suggest_friends(self, user_id, text):
     return suggest_friends(user_id, text)
 
 
-@celery.task(name="async_get_friend_invites", bind=True, base=SqlAlchemyTask)
+@celery.task(name="async_get_friend_invites", bind=True, base=BaseTask)
 def async_get_friend_invites(self, user_id):
     invite_ids = get_friend_invite_ids(user_id)
     if not invite_ids:
@@ -322,28 +323,28 @@ endpoint in order to have user data be as dynamic and responsive as possible:
 """
 
 
-@celery.task(name="async_serialize_open_orders", bind=True, base=SqlAlchemyTask)
+@celery.task(name="async_serialize_open_orders", bind=True, base=BaseTask)
 def async_serialize_open_orders(self, game_id, user_id):
     serialize_and_pack_orders_open_orders(game_id, user_id)
 
 
-@celery.task(name="async_serialize_current_balances", bind=True, base=SqlAlchemyTask)
+@celery.task(name="async_serialize_current_balances", bind=True, base=BaseTask)
 def async_serialize_current_balances(self, game_id, user_id):
     serialize_and_pack_current_balances(game_id, user_id)
 
 
-@celery.task(name="async_serialize_balances_chart", bind=True, base=SqlAlchemyTask)
+@celery.task(name="async_serialize_balances_chart", bind=True, base=BaseTask)
 def async_serialize_balances_chart(self, game_id, user_id):
     df = make_balances_chart_data(game_id, user_id)
     serialize_and_pack_balances_chart(df, game_id, user_id)
 
 
-@celery.task(name="async_make_the_field_charts", bind=True, base=SqlAlchemyTask)
+@celery.task(name="async_make_the_field_charts", bind=True, base=BaseTask)
 def async_make_the_field_charts(self, game_id):
     make_the_field_charts(game_id)
 
 
-@celery.task(name="async_update_play_game_visuals", bind=True, base=SqlAlchemyTask)
+@celery.task(name="async_update_play_game_visuals", bind=True, base=BaseTask)
 def async_update_play_game_visuals(self):
     open_game_ids = get_active_game_ids()
     task_results = list()
@@ -356,7 +357,7 @@ def async_update_play_game_visuals(self):
     pause_return_until_subtask_completion(task_results, "async_update_play_game_visuals")
 
 
-@celery.task(name="async_calculate_metrics", bind=True, base=SqlAlchemyTask)
+@celery.task(name="async_calculate_metrics", bind=True, base=BaseTask)
 def async_calculate_game_metrics(self, game_id, user_id, start_date=None, end_date=None):
     calculate_and_pack_metrics(game_id, user_id, start_date, end_date)
 
@@ -365,7 +366,7 @@ def async_calculate_game_metrics(self, game_id, user_id, start_date=None, end_da
 # Player stat production #
 # ---------------------- #
 
-@celery.task(name="async_update_player_stats", bind=True, base=SqlAlchemyTask)
+@celery.task(name="async_update_player_stats", bind=True, base=BaseTask)
 def async_update_player_stats(self):
     """This task calculates game-level metrics for all players in all games, caching those metrics to redis
     """
@@ -378,11 +379,11 @@ def async_update_player_stats(self):
     pause_return_until_subtask_completion(task_results, "async_update_player_stats")
 
 
-@celery.task(name="async_compile_player_stats", bind=True, base=SqlAlchemyTask)
+@celery.task(name="async_compile_player_stats", bind=True, base=BaseTask)
 def async_compile_player_sidebar_stats(self, game_id):
     compile_and_pack_player_sidebar_stats(game_id)
 
 
-@celery.task(name="async_get_player_cash_balance", bind=True, base=SqlAlchemyTask)
+@celery.task(name="async_get_player_cash_balance", bind=True, base=BaseTask)
 def async_get_player_cash_balance(self, game_id, user_id):
     return get_current_game_cash_balance(user_id, game_id)
